+++ qiskit_experiments/calibration_management/__init__.py
@@ -36,7 +36,6 @@ Classes
 .. autosummary::
     :toctree: ../stubs/
 
-    BackendCalibrations
     Calibrations
 
 Managing Calibration Data
@@ -144,6 +143,5 @@ which would return the default xp schedule block template for all qubits.
 """
 
 from .calibrations import Calibrations
-from .backend_calibrations import BackendCalibrations
 from .base_calibration_experiment import BaseCalibrationExperiment
 from .basis_gate_library import FixedFrequencyTransmon
