+++ qiskit_experiments/library/calibration/half_angle_cal.py
@@ -20,7 +20,7 @@ from qiskit.providers.backend import Backend
 from qiskit_experiments.framework import ExperimentData
 from qiskit_experiments.calibration_management import (
     BaseCalibrationExperiment,
-    BackendCalibrations,
+    Calibrations,
 )
 from qiskit_experiments.library.characterization import HalfAngle
 from qiskit_experiments.calibration_management.update_library import BaseUpdater
@@ -32,7 +32,7 @@ class HalfAngleCal(BaseCalibrationExperiment, HalfAngle):
     def __init__(
         self,
         qubit,
-        calibrations: BackendCalibrations,
+        calibrations: Calibrations,
         backend: Optional[Backend] = None,
         schedule_name: str = "sx",
         cal_parameter_name: Optional[str] = "amp",
