+++ qiskit_experiments/library/calibration/rough_amplitude_cal.py
@@ -21,7 +21,7 @@ from qiskit.circuit import Parameter
 from qiskit.providers.backend import Backend
 
 from qiskit_experiments.framework import ExperimentData, Options
-from qiskit_experiments.calibration_management import BaseCalibrationExperiment, BackendCalibrations
+from qiskit_experiments.calibration_management import BaseCalibrationExperiment, Calibrations
 from qiskit_experiments.library.characterization import Rabi
 from qiskit_experiments.calibration_management.update_library import BaseUpdater
 from qiskit_experiments.curve_analysis import ParameterRepr
@@ -41,7 +41,7 @@ class RoughAmplitudeCal(BaseCalibrationExperiment, Rabi):
     def __init__(
         self,
         qubit: int,
-        calibrations: BackendCalibrations,
+        calibrations: Calibrations,
         schedule_name: str = "x",
         amplitudes: Iterable[float] = None,
         cal_parameter_name: Optional[str] = "amp",
@@ -196,7 +196,7 @@ class RoughXSXAmplitudeCal(RoughAmplitudeCal):
     def __init__(
         self,
         qubit: int,
-        calibrations: BackendCalibrations,
+        calibrations: Calibrations,
         amplitudes: Iterable[float] = None,
         backend: Optional[Backend] = None,
     ):
@@ -225,7 +225,7 @@ class EFRoughXSXAmplitudeCal(RoughAmplitudeCal):
     def __init__(
         self,
         qubit: int,
-        calibrations: BackendCalibrations,
+        calibrations: Calibrations,
         amplitudes: Iterable[float] = None,
         backend: Optional[Backend] = None,
         ef_pulse_label: str = "12",
