+++ qiskit_experiments/library/calibration/rough_frequency.py
@@ -18,7 +18,7 @@ from qiskit.providers.backend import Backend
 from qiskit_experiments.library.characterization.qubit_spectroscopy import QubitSpectroscopy
 from qiskit_experiments.library.characterization.ef_spectroscopy import EFSpectroscopy
 from qiskit_experiments.calibration_management.update_library import Frequency
-from qiskit_experiments.calibration_management.backend_calibrations import BackendCalibrations
+from qiskit_experiments.calibration_management.calibrations import Calibrations
 from qiskit_experiments.calibration_management.base_calibration_experiment import (
     BaseCalibrationExperiment,
 )
@@ -30,7 +30,7 @@ class RoughFrequencyCal(BaseCalibrationExperiment, QubitSpectroscopy):
     def __init__(
         self,
         qubit: int,
-        calibrations: BackendCalibrations,
+        calibrations: Calibrations,
         frequencies: Iterable[float],
         backend: Optional[Backend] = None,
         unit: str = "Hz",
@@ -77,7 +77,7 @@ class RoughEFFrequencyCal(BaseCalibrationExperiment, EFSpectroscopy):
     def __init__(
         self,
         qubit: int,
-        calibrations: BackendCalibrations,
+        calibrations: Calibrations,
         frequencies: Iterable[float],
         unit: str = "Hz",
         auto_update: bool = True,
