+++ test/calibration/experiments/test_fine_amplitude.py
@@ -30,7 +30,7 @@ from qiskit_experiments.library import (
     FineSXAmplitudeCal,
 )
 from qiskit_experiments.calibration_management.basis_gate_library import FixedFrequencyTransmon
-from qiskit_experiments.calibration_management import BackendCalibrations
+from qiskit_experiments.calibration_management import Calibrations
 from qiskit_experiments.test.mock_iq_backend import MockFineAmp
 
 
@@ -158,7 +158,7 @@ class TestFineAmplitudeCal(QiskitExperimentsTestCase):
         library = FixedFrequencyTransmon()
 
         self.backend = MockFineAmp(-np.pi * 0.07, np.pi, "xp")
-        self.cals = BackendCalibrations(self.backend, library)
+        self.cals = Calibrations.from_backend(self.backend, library)
 
     def test_cal_options(self):
         """Test that the options are properly propagated."""
